+++ automotive/src/main/java/de/moleman1024/audiowagon/filestorage/usb/USBAudioDataSource.kt
@@ -12,6 +12,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.runBlocking
 import java.io.IOException
 import java.nio.ByteBuffer
+import java.util.concurrent.TimeoutException
 import kotlin.math.min
 
 private const val TAG = "USBAudDataSrc"
@@ -32,10 +33,16 @@ open class USBAudioDataSource(
     var hasError = false
 
     init {
-        runBlocking(libaumsDispatcher) {
-            logger.verbose(
-                TAG, "Init data source with chunk size $chunkSize for file with length ${usbFile?.length}: $usbFile"
-            )
+        try {
+            runBlocking(libaumsDispatcher) {
+                logger.verbose(
+                    TAG, "Init data source with chunk size $chunkSize for file with length ${usbFile?.length}: $usbFile"
+                )
+            }
+        } catch (exc: TimeoutException) {
+            logger.exception(TAG, exc.message.toString(), exc)
+            hasError = true
+            throw IOException(exc.message.toString())
         }
     }
 
@@ -44,16 +51,21 @@ open class USBAudioDataSource(
             usbFile = null
             return
         }
-        runBlocking(libaumsDispatcher) {
-            try {
-                usbFile?.close()
-                isClosed = true
-                usbFile = null
-            } catch (exc: IOException) {
-                // If we rethrow this, it will trigger a fatal SIGABRT and the app will be restarted.
-                logger.exceptionLogcatOnly(TAG, "Error while closing USB file, did you unplug the device?", exc)
-                hasError = true
+        try {
+            runBlocking(libaumsDispatcher) {
+                try {
+                    usbFile?.close()
+                    isClosed = true
+                    usbFile = null
+                } catch (exc: IOException) {
+                    // If we rethrow this, it will trigger a fatal SIGABRT and the app will be restarted.
+                    logger.exceptionLogcatOnly(TAG, "Error while closing USB file, did you unplug the device?", exc)
+                    hasError = true
+                }
             }
+        } catch (exc: TimeoutException) {
+            logger.exception(TAG, exc.message.toString(), exc)
+            hasError = true
         }
     }
 
@@ -109,8 +121,14 @@ open class USBAudioDataSource(
     }
 
     override fun getSize(): Long {
-        return runBlocking(libaumsDispatcher) {
-            return@runBlocking usbFile?.length ?: 0
+        try {
+            return runBlocking(libaumsDispatcher) {
+                return@runBlocking usbFile?.length ?: 0
+            }
+        } catch (exc: TimeoutException) {
+            logger.exception(TAG, exc.message.toString(), exc)
+            hasError = true
+            throw IOException(exc.message.toString())
         }
     }
 
@@ -131,6 +149,10 @@ open class USBAudioDataSource(
             logger.exception(TAG, exc.message.toString(), exc)
             hasError = true
             throw exc
+        } catch (exc: TimeoutException) {
+            logger.exception(TAG, exc.message.toString(), exc)
+            hasError = true
+            throw IOException(exc.message.toString())
         }
     }