+++ automotive/src/main/java/de/moleman1024/audiowagon/log/LogBuffer.kt
@@ -76,7 +76,7 @@ class LogBuffer(private val numMaxEntries: Int = NUM_MAX_LOG_ENTRIES) {
                 }
                 val logData = iterator.previous()
                 if (logData.stackTrace?.isBlank() == false) {
-                    logData.stackTrace.lines().reversed().forEach { stackTraceLine ->
+                    logData.stackTrace.lines().forEach { stackTraceLine ->
                         if (stackTraceLine.isNotBlank()) {
                             logData.msg = stackTraceLine
                             val formattedStackTraceLine = formatLogData(logData)